requirements.txt
langchain==0.2.3
langchain-community==0.2.3
langchain-google-genai==1.0.10
google-generativeai==0.7.0
python-dotenv==1.0.1
nba_api==1.1.13
pandas==2.2.2
requests==2.31.0
beautifulsoup4==4.12.3
streamlit==1.35.0
selenium
webdriver-manager
pytz